Rewrote the code because I was ashamed of it
So ashamed that I beg you to don't look the older commits (?
Btw, the changes here basically are that I rewrote some stuff in a nicer way.
Also I'm now following 'properly' the PEP8. But since It tells me to indent breaking the 72 characters rule (that is
way more important in my opinion) I decided to ignore it.