Reorganized
The "oldsrc" folder will be deleted by the end of my wonderful job on
this branch. This is another "rewrite" of the code. Because now I have
a clearer idea of what I want to do.

So, explaining a little:
src > Where the code will be.
res > Where the resources will be, at this moment, just the GUI.
oldsrc > Where the old code is. To be deleted.
docs > Where the documents someday will be.
tests > Where the testsuits will be.
examples > I have a dream that someday I will use this.